chapter-3-small-project/pyproject.toml
[project]
name = "web-crawler"
version = "0.1.0"
description = "A real-time web crawler with search capabilities for extracting and parsing web content"
readme = "README.md"
authors = [
    { name = "virattt", email = "support@financialdatasets.ai" }
]
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.15",
    "pydantic>=2.8.2",
    "beautifulsoup4>=4.13.4",
    "googlenewsdecoder>=0.1.7",
    "playwright>=1.54.0",
    "requests>=2.32.4",
    "feedparser>=6.0.11",
]

[project.scripts]
web-crawler = "src.main:main"
chapter3-small-project = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23"
]
